data/requirements.txt
espn-api>=0.45.0
python-dotenv>=1.0.0
